mini-application for a banking system in Java
mini-application for a banking system in Java. In this program, we will add some basic functionalities of a bank account like a deposit of amount, withdrawal of amount, etc.Initially, the program accepts the number of customers (Array of Objects)we need to add and adds the customer and account details accordingly. Further, it displays the series of menus to operate over the accounts.

1.Display all account details

2.Search by account number

3.Deposit the amount

4.Withdraw the amount

5.Exit 

Create Custom Exception and perform the following action:

Customers are not allowed to deposit amount <= 0 ( In this case throw InvalidAmountException).

Customers are not allowed to withdraw amount <= 0 (throw InvalidAmountException).

Customers are also not allowed to withdraw an amount greater than (>) the available amount (throw InsufficientFundsException)